+++ pkg/parser/image/parser_test.go
@@ -1,4 +1,4 @@
-package parser
+package image
 
 import (
 	"testing"
